Move remote upstream capabilities to subpackage
