+++ dependency/dependency.go
@@ -27,12 +27,8 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
-	"github.com/mitchellh/mapstructure"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v3"
-
-	"sigs.k8s.io/zeitgeist/upstream"
 )
 
 // Client holds any client that is needed
@@ -133,7 +129,7 @@ func (decoded *Dependency) UnmarshalYAML(unmarshal func(interface{}) error) erro
 	return nil
 }
 
-func fromFile(dependencyFilePath string) (*Dependencies, error) {
+func FromFile(dependencyFilePath string) (*Dependencies, error) {
 	depFile, err := os.ReadFile(dependencyFilePath)
 	if err != nil {
 		return nil, err
@@ -149,7 +145,7 @@ func fromFile(dependencyFilePath string) (*Dependencies, error) {
 	return dependencies, nil
 }
 
-func toFile(dependencyFilePath string, dependencies *Dependencies) error {
+func ToFile(dependencyFilePath string, dependencies *Dependencies) error {
 	var output bytes.Buffer
 	yamlEncoder := yaml.NewEncoder(&output)
 	yamlEncoder.SetIndent(2)
@@ -179,7 +175,7 @@ func NewLocalClient() (Client, error) {
 // Will return an error if the dependency cannot be found in the files it has defined, or if the version does not match
 func (c *LocalClient) LocalCheck(dependencyFilePath, basePath string) error {
 	log.Debugf("Base path: %s", basePath)
-	externalDeps, err := fromFile(dependencyFilePath)
+	externalDeps, err := FromFile(dependencyFilePath)
 	if err != nil {
 		return err
 	}
@@ -263,392 +259,6 @@ func (c *LocalClient) RemoteExport(dependencyFilePath string) ([]VersionUpdate, 
 	return nil, UnsupportedError{"remote export is not supported by the local client"}
 }
 
-type RemoteClient struct {
-	LocalClient  Client
-	AWSEC2Client ec2iface.EC2API
-}
-
-func NewRemoteClient() (Client, error) {
-	localClient, err := NewLocalClient()
-	if err != nil {
-		return nil, err
-	}
-	return &RemoteClient{
-		LocalClient:  localClient,
-		AWSEC2Client: upstream.NewAWSClient(),
-	}, nil
-}
-
-func (c *RemoteClient) LocalCheck(dependencyFilePath, basePath string) error {
-	return c.LocalClient.LocalCheck(dependencyFilePath, basePath)
-}
-
-// RemoteCheck checks whether dependencies are up to date with upstream
-//
-// Will return an error if checking the versions upstream fails.
-//
-// Out-of-date dependencies will be printed out on stdout at the INFO level.
-func (c *RemoteClient) RemoteCheck(dependencyFilePath string) ([]string, error) {
-	externalDeps, err := fromFile(dependencyFilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	updates := make([]string, 0)
-
-	versionUpdateInfos, err := c.checkUpstreamVersions(externalDeps.Dependencies)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, vu := range versionUpdateInfos {
-		if vu.updateAvailable {
-			updates = append(
-				updates,
-				fmt.Sprintf(
-					"Update available for dependency %s: %s (current: %s)",
-					vu.name,
-					vu.latest.Version,
-					vu.current.Version,
-				),
-			)
-		} else {
-			log.Debugf(
-				"No update available for dependency %s: %s (latest: %s)\n",
-				vu.name,
-				vu.current.Version,
-				vu.latest.Version,
-			)
-		}
-	}
-
-	return updates, nil
-}
-
-// Upgrade retrieves the most up-to-date version of the dependency and replaces
-// the local version with the most up-to-date version.
-//
-// Will return an error if checking the versions upstream fails, or if updating
-// files fails.
-func (c *RemoteClient) Upgrade(dependencyFilePath, basePath string) ([]string, error) {
-	externalDeps, err := fromFile(dependencyFilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	upgrades := make([]string, 0)
-	upgradedDependencies := make([]*Dependency, 0)
-
-	versionUpdateInfos, err := c.checkUpstreamVersions(externalDeps.Dependencies)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, vu := range versionUpdateInfos {
-		dependency, err := findDependencyByName(externalDeps.Dependencies, vu.name)
-		if err != nil {
-			return nil, err
-		}
-
-		if vu.updateAvailable {
-			err = upgradeDependency(basePath, dependency, &vu)
-			if err != nil {
-				return nil, err
-			}
-
-			dependency.Version = vu.latest.Version
-			upgradedDependencies = append(
-				upgradedDependencies,
-				dependency,
-			)
-
-			upgrades = append(
-				upgrades,
-				fmt.Sprintf(
-					"Upgraded dependency %s from version %s to version %s",
-					vu.name,
-					vu.current.Version,
-					vu.latest.Version,
-				),
-			)
-		} else {
-			upgradedDependencies = append(
-				upgradedDependencies,
-				dependency,
-			)
-
-			log.Debugf(
-				"No update available for dependency %s: %s (latest: %s)\n",
-				vu.name,
-				vu.current.Version,
-				vu.latest.Version,
-			)
-		}
-	}
-
-	// Update the dependencies file to reflect the upgrades
-	err = toFile(dependencyFilePath, &Dependencies{
-		Dependencies: upgradedDependencies,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return upgrades, nil
-}
-
-// SetVersion sets the version of a dependency to the specified version
-//
-// Will return an error  if updating files fails.
-func (c *Client) SetVersion(dependencyFilePath, basePath, dependency, version string) error {
-	externalDeps, err := fromFile(dependencyFilePath)
-	if err != nil {
-		return err
-	}
-
-	found := false
-	for _, dep := range externalDeps.Dependencies {
-		if dep.Name == dependency {
-			found = true
-
-			if err := upgradeDependency(basePath, dep, &versionUpdateInfo{
-				name: dep.Name,
-				current: Version{
-					Version: dep.Version,
-					Scheme:  dep.Scheme,
-				},
-				latest: Version{
-					Version: version,
-					Scheme:  dep.Scheme,
-				},
-				updateAvailable: true,
-			}); err != nil {
-				return err
-			}
-
-			dep.Version = version
-		}
-	}
-
-	if !found {
-		return fmt.Errorf("dependency %s not found", dependency)
-	}
-
-	// Update the dependencies file to reflect the upgrades
-	err = toFile(dependencyFilePath, externalDeps)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func findDependencyByName(dependencies []*Dependency, name string) (*Dependency, error) {
-	for _, dep := range dependencies {
-		if dep.Name == name {
-			return dep, nil
-		}
-	}
-	return nil, fmt.Errorf("cannot find dependency by name: %s", name)
-}
-
-func upgradeDependency(basePath string, dependency *Dependency, versionUpdate *versionUpdateInfo) error {
-	log.Debugf("running upgradeDependency, versionUpdate %#v", versionUpdate)
-	for _, refPath := range dependency.RefPaths {
-		err := replaceInFile(basePath, refPath, versionUpdate)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func replaceInFile(basePath string, refPath *RefPath, versionUpdate *versionUpdateInfo) error {
-	filename := filepath.Join(basePath, refPath.Path)
-	log.Debugf("running replaceInFile, refpath is %#v, versionUpdate %#v", refPath, versionUpdate)
-
-	matcher, err := regexp.Compile(refPath.Match)
-	if err != nil {
-		return fmt.Errorf("compiling regex: %w", err)
-	}
-
-	inputFile, err := os.ReadFile(filename)
-	if err != nil {
-		return fmt.Errorf("reading file: %w", err)
-	}
-
-	lines := strings.Split(string(inputFile), "\n")
-
-	for i, line := range lines {
-		if matcher.MatchString(line) {
-			if strings.Contains(line, versionUpdate.current.Version) {
-				log.Debugf(
-					"Line %d matches expected regexp %q and version %q: %s",
-					i,
-					refPath.Match,
-					versionUpdate.current.Version,
-					line,
-				)
-
-				// The actual upgrade:
-				lines[i] = strings.ReplaceAll(line, versionUpdate.current.Version, versionUpdate.latest.Version)
-			}
-		}
-	}
-
-	upgradedFile := strings.Join(lines, "\n")
-
-	// Finally, write the file out
-	err = os.WriteFile(filename, []byte(upgradedFile), 0o644)
-
-	if err != nil {
-		return fmt.Errorf("writing file: %w", err)
-	}
-	return nil
-}
-
-func (c *RemoteClient) RemoteExport(dependencyFilePath string) ([]VersionUpdate, error) {
-	externalDeps, err := fromFile(dependencyFilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	versionUpdates := []VersionUpdate{}
-
-	versionUpdatesInfos, err := c.checkUpstreamVersions(externalDeps.Dependencies)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, vui := range versionUpdatesInfos {
-		if vui.updateAvailable {
-			versionUpdates = append(versionUpdates, VersionUpdate{
-				Name:       vui.name,
-				Version:    vui.current.Version,
-				NewVersion: vui.latest.Version,
-			})
-		} else {
-			log.Debugf(
-				"No update available for dependency %s: %s (latest: %s)\n",
-				vui.name,
-				vui.current.Version,
-				vui.latest.Version,
-			)
-		}
-	}
-	return versionUpdates, nil
-}
-
-func (c *RemoteClient) checkUpstreamVersions(deps []*Dependency) ([]versionUpdateInfo, error) {
-	versionUpdates := []versionUpdateInfo{}
-	for _, dep := range deps {
-		if dep.Upstream == nil {
-			versionUpdates = append(versionUpdates, versionUpdateInfo{
-				name:            dep.Name,
-				current:         Version{dep.Version, dep.Scheme},
-				updateAvailable: false,
-			})
-			continue
-		}
-
-		up := dep.Upstream
-		latestVersion := Version{dep.Version, dep.Scheme}
-		currentVersion := Version{dep.Version, dep.Scheme}
-
-		var err error
-
-		// Cast the flavour from the currently unknown upstream type
-		flavour := upstream.Flavour(up["flavour"])
-		switch flavour {
-		case upstream.DummyFlavour:
-			var d upstream.Dummy
-
-			decodeErr := mapstructure.Decode(up, &d)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = d.LatestVersion()
-		case upstream.GithubFlavour:
-			var gh upstream.Github
-
-			decodeErr := mapstructure.Decode(up, &gh)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = gh.LatestVersion()
-		case upstream.GitLabFlavour:
-			var gl upstream.GitLab
-
-			decodeErr := mapstructure.Decode(up, &gl)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = gl.LatestVersion()
-		case upstream.HelmFlavour:
-			var h upstream.Helm
-
-			decodeErr := mapstructure.Decode(up, &h)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = h.LatestVersion()
-		case upstream.AMIFlavour:
-			var ami upstream.AMI
-
-			decodeErr := mapstructure.Decode(up, &ami)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			ami.ServiceClient = c.AWSEC2Client
-
-			latestVersion.Version, err = ami.LatestVersion()
-		case upstream.ContainerFlavour:
-			var ct upstream.Container
-
-			decodeErr := mapstructure.Decode(up, &ct)
-			if decodeErr != nil {
-				log.Debug("errr decoding")
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = ct.LatestVersion()
-		case upstream.EKSFlavour:
-			var eks upstream.EKS
-
-			decodeErr := mapstructure.Decode(up, &eks)
-			if decodeErr != nil {
-				return nil, decodeErr
-			}
-
-			latestVersion.Version, err = eks.LatestVersion()
-		default:
-			return nil, fmt.Errorf("unknown upstream flavour '%#v' for dependency %s", flavour, dep.Name)
-		}
-
-		if err != nil {
-			return nil, err
-		}
-
-		updateAvailable, err := latestVersion.MoreSensitivelyRecentThan(currentVersion, dep.Sensitivity)
-		if err != nil {
-			return nil, err
-		}
-
-		latestVersion.Version = formatVersion(currentVersion.Version, latestVersion.Version)
-
-		versionUpdates = append(versionUpdates, versionUpdateInfo{
-			name:            dep.Name,
-			current:         currentVersion,
-			latest:          latestVersion,
-			updateAvailable: updateAvailable,
-		})
-	}
-
-	return versionUpdates, nil
+var NewRemoteClient = func() (Client, error) {
+	return nil, UnsupportedError{"remote upstream functionality is not supported by this command; use sigs.k8s.io/zeitgeist/remote/zeitgeist"}
 }
