+++ remote/testdata/remote-ami.yaml
